fix: Triggering Github Release
After publish changelog log and there are changes worth making a new tag
for, a release should be triggered. This code was commented out during
testing of a new way to trigger a tag if you use CircleCI's new dynamic
workflows. This commit uncomments the code and places it behind a flag
for backwards compatibility it is default to true.