add: Deley After Publishing Changelog
Added 3 seconds after the changelog updatee pull request is merged into
main, and before the new workflow is triggered. This should ensure that
by the time the tag-and-release workflow runs and it does it checks to
see if the changelog needs update, it will not and succeed. If it does
then it would mean another merge with taggable changes is running, in
which case it should fail. Added a 3 second delay to the start of the
publich changelog workflow to see if that helps this problem as well.