chg: Over To Use Version Release Orb CLI
Integrate the Version Release Orb CLI tool to replace bash scripts.
All this is in an effort to fix bugs causing workflows to trigger
unexpectidly. Causing confusion as to what was happening. Making
the workflow-selector the only workflow that gets triggered when code is
merged into the trunk branch. This allows more precise control as to
what is triggerd and how, replacing the confusing distributed model.

The following changes may also be breaking changes:

* Added variable to toggle attachment of the workspace before any other
  steps in the publish-changelog job.
* Removed/added/renamed job parameters.
* Removed publish-changelog.sh script.
* Updated documentation to reflect the changes in the interface.
* Removed txt file use to indicate to the publish-changelog job that it
  should trigger a release tag workflow.
* Added CI job to publish container image for version release, which
  replaces the git-tool-belt image that was the default executor.
* Changed workflows to trigger when trigger_flow parameter is set.
* Added a custom mock-server to run test that require a remote server
  response.
* Set which version of the Version Release orb to use.

BREAKING CHANGE

* Removed jobs, commands, and examples along with their companion scripts.