fix: Release Image
Docker COPY command was missing the --from argument. This caused the
image to be overwritten with everything from the scratch image itself.
Not sure why that is even possible. But copy does not check to see if
the source is the same as the destination, if it does, then it does not
warn the user.