BREAKING CHANGE!: Replace vro with avr Command
The `avr` command is a rewrite of `vro`. While the
tool functions mostly the same it is not backwards compatible because of
how command line options and arguments are passed in. This was to
accomodate the support of GitHub actions. Code that could be reused for
both CircleCI and GitHub was modified to handle arguments to accomodate
both. While the command line tool has changed, the CircleCI Orb was
mostly unaffected, with the exception that deprecated options have been
removed.

BREAKING CHANGE