chg: Change GitHub Token Environment Variable Name
This changes the name of the GitHub token
environment variable used in the CircleCI
workflows. It is now consistent with the name of
the token used for GitHub Actions. However, due to
fundamental differences in how CircleCI and GitHub
Actions work, this change will require
developers to directly pass in the token to each
workflow job that requires it.

BREAKING CHANGE: You are now required to pass the token in directly to a workflow.