chg: Push Commit Through GitHub REST API
To beable to push with using a fine-grained access to it mains
changing from using the Git CLI to using the GitHub REST API.
This commit adds:

* Added functionality to make a new tree of changes using the GitHub
  REST API.
* Added functionality to make a new commit from a tree of changes using
  the GitHub REST API.
* Added functionality to get a reference from the GitHub REST API.
* Added functionality to make a new branch using an existing reference
  as a base through the GitHub REST API.
* Added functionality to point a branch to a commit using the GitHub REST API.

BREAKING CHANGE: There is a fair amount of code that was replaced or
removed. This is a breaking change.

History

With CircleCI we used an SSH key/pair to allow write to the GitHub
repository. We can use that again, but it would be less complicated
to use a GitHub Fine-grained access token. This will work for the
new GitHub Actions implementation. While also allowing us to remove
the SSH key/pair from the CircleCI implementation and have it continue
to work as it does today. That is a nice trade off and benefit.