fix: Publish Changelog During Auto Release
The changelog had the full working directory path, this caused
issues with the release process. This commit removes the full path
and restores it to the path that is relative to the repository it
is releasing. This change has also been made for the changelog
generator configuration (Git-Cliff), which should have suffered
from the same problem.