chg: Refactor Adding Files to Changelog Commit
After some thought, the new way to add files to the changelog commit
is more robust than required. While initially the though was to have
git just add any changes files seem well suited, this boils down to
blindly adding any genearated files to the commit. The next
implementation will allow the user to specify a file which will
contain a list of files to add to the commit. This gives the client
much greater control without bloating the CI configuration. Which
is why the initial decision to allow the user to specify a a list
of file in the CI was never implemented. Insteadn this overly
complex solution was implemented. However, this was aborted before
fullly implementing the solution. The next implementation will
be much simpler and more robust.