chg: Add Additional Commit Files
Adding additional files for CircleCI and GitHub Actions works
different with the only simlarities being they use artifacts to
pass files from job to downstream jobs.

With GitHub Actions you must use the upload-artifact action.
Build a directory that will be used as an artifact containting
all the files you wish to commit, and a file that serves as a
manifest and lists those same files. Then point upload-artifact
path parameter to the directory, which will be zipped and
uploaded. The download-artifact job runs just before
publish-changelog, will download the uploaded zip and
extract all files to the root of the repository; without the
parent directory itself.

With CircleCI you can add a job just before publish-changelong,
build any files and generate a manifest file listing each of
those files. Pass the filename of the manifest that list the
additional files to publish-changelog job.

The publish-changelog job, in either case will use the file that
serves as a manifest and lists additonal files to add to the
commit; along with the changelog and the git-cliff configuration
file if it is missing.