fix: Configure GitHub Server and API
A bug was found when passing in the GitHub server domain to the
`avr` application. It was never set. Also the API was set directly
via environment variable, making a different way to set a simlar
variable. In addition, the GitHub API URL was set as GITHUB_API_URL
against GitHub Actions recommendation of using GITHUB as a prefix.
It will be changed to PARAM_GITHUB_API_URL. Since we have just
released avr 5.0.0, we will mark it as a fix instead of a breaking
change. Also GitHub Server domain and GitHub API URL are now set
as global options since they are used in multiple workflows.

This commit fixes setting GitHub Server domain and GitHub API URL,
so that they are done in a consistent way. By passing in an
option to th avr appliation.