fix: Permissions Required for Publish Changelog
It would be nice to specify the Permissions that each job requires,
however, it has been observed that the `GITHUB_TOKEN` does not
have the required permissions to publish the changelog. Even when set
directly, until this is figured out, we'll have to settle for leaving
comments in the workflow file.